requirements.txt
flask==2.0.2
flask-sqlalchemy==2.5.1
flask-restful==0.3.9
SQLAlchemy-serializer==1.3.4.2
pymysql==1.0.2
ipython==7.31.0
python-dotenv
sphinx_rtd_theme
sphinxcontrib-openapi
sphinxcontrib-redoc